+++ HISTORY.rst
@@ -25,6 +25,7 @@ Latest (unreleased)
 * Install sale_contract
 * Remove signatures on leads
 * update transitions checks in crm
+* Add NDA Leads in customer's form
 
 **Bugfixes**